Create Help Martha.cpp
Martha, a girl has been provided with its starting co-ordinate on a 2-D plane (x,y). She is to provided with a string that tells her to move in four cardinal directions namely up(U), down(D), left(L) and right(R). She has two choices - either take that direction in the string(S) or leave it. The problem wants to output the minimum number of steps to reach the targets if there is any way to reach the destinations following a sub-sequence of the string.

Input-
1                      #Testcases
RLUDRLUD      #string of directions
1 2                   #Start co-ordinate
2                      #number of targets
2 3                   #1st target
1 -2                  #2nd target

Output-
YES 2
NO