Create Election.cpp
There is a queue of voters from left to right. The election is held between two candidates 'A' and 'B'. People standing are either voting A or B and some are neutral that they have not decided yet, denoted as '-'. A person standing in the queue, if voting A or B will try to convince neutral people in the queue to vote their candidate. People supporting A will convince the people at their left and for B, they will approach to the right. The neutral candidate vote in the favor of those who reaches him first. 

After the whole process, whosoever gets maximum number of people voting wins.
In the case of equal votes, the result will be a "Coalition government".

Input-
14                            #length of string
--A--BB---AB--        #string representing queue

Output-
B